refactor(fixture): rename fixture to ts-fixture
