+++ package.json
@@ -12,7 +12,7 @@
     "karma-debug": "cross-env ./node_modules/.bin/nps karma.watch.debug",
     "build-ioc": "rollup -c ioc.rollup.config.js",
     "test-ioc-src": "node ioc/test/generate src",
-    "test-ioc-fixture": "node ioc/test/generate test/fixture",
+    "test-ioc-ts-fixture": "node ioc/test/generate test/fixture/ts",
     "transpile": "tsc src/compile.ts --outDir scripts",
     "build": "rollup -c compiler.rollup.config.js",
     "test": "node compiler/test/compile",
