+++ test/fixture/ts/components/article/articlecomponent.html
