+++ test/fixture/ts/components/article/articlecomponent.ts
