+++ test/fixture/ts/components/article/comment.html
