+++ test/fixture/ts/components/article/comment.ts
