+++ test/fixture/ts/components/auth/authcomponent.html
