+++ test/fixture/ts/components/auth/authcomponent.ts
