+++ test/fixture/ts/components/editor/editorcomponent.html
