+++ test/fixture/ts/components/editor/editorcomponent.ts
