+++ test/fixture/ts/components/home/homecomponent.html
