+++ test/fixture/ts/components/home/homecomponent.ts
