+++ test/fixture/ts/components/profile/profilearticlecomponent.html
