+++ test/fixture/ts/components/profile/profilearticlecomponent.ts
