+++ test/fixture/ts/components/profile/profilecomponent.html
