+++ test/fixture/ts/components/profile/profilecomponent.ts
