+++ test/fixture/ts/components/profile/profilefavoritescomponent.html
