+++ test/fixture/ts/components/profile/profilefavoritescomponent.ts
