+++ test/fixture/ts/components/settings/settingscomponent.html
