+++ test/fixture/ts/components/settings/settingscomponent.ts
