+++ test/fixture/ts/environment.ts
