+++ test/fixture/ts/generate-raw.js
