+++ test/fixture/ts/globals.d.ts
