+++ test/fixture/ts/index.ts
