+++ test/fixture/ts/main.ts
