+++ test/fixture/ts/raw.ts
