+++ test/fixture/ts/resources/elements/articlelist.ts
