+++ test/fixture/ts/resources/elements/articlepreview.ts
