+++ test/fixture/ts/resources/index.ts
