+++ test/fixture/ts/resources/value-converters/keys.ts
