+++ test/fixture/ts/shared/layouts/footerlayout.ts
