+++ test/fixture/ts/shared/models/user.ts
