+++ test/fixture/ts/shared/services/config.ts
