+++ test/fixture/ts/shared/state/sharedstate.ts
