+++ test/unit/ioc/injector-builder.spec.ts
@@ -36,12 +36,12 @@ import {
   ProfileFavoritesComponent,
   SettingsComponent,
   UserService
-} from '../../fixture';
+} from '../../fixture/ts';
 import { InjectorBuilder, DefaultInjector } from '../../../ioc/injector';
 import { DependencyType } from '../../../ioc/types';
 import { IContext } from '../../../ioc/interfaces';
 import * as ts from 'typescript';
-import { raw } from '../../fixture/raw';
+import { raw } from '../../fixture/ts/raw';
 import * as AST from '../../../ioc/analysis/ast';
 import { TypeScriptSyntaxTransformer } from '../../../ioc/analysis/ts-syntax-transformer';
